Fix Int8Projection: validated deserialize + predict returns usize (GPT review)
Two P1 bugs fixed:
1. Deserialize now validates weights.len() == input_dim * output_classes
   via private ProjectionDisk DTO (same pattern as Network genome).
   Malformed payloads are rejected instead of causing predict() panics.
2. predict() returns usize instead of u8, preventing class index
   truncation above 255 classes.

2 regression tests added. 129 tests pass, clippy clean.

+++ instnct-core/examples/evolve_language.rs
@@ -122,7 +122,7 @@ fn eval_accuracy(
     let mut correct = 0u32;
     for i in 0..len {
         net.propagate(&sdr[seg[i] as usize], config).unwrap();
-        if projection.predict(&net.charge()[OUTPUT_START..NEURON_COUNT]) == seg[i + 1] {
+        if projection.predict(&net.charge()[OUTPUT_START..NEURON_COUNT]) == seg[i + 1] as usize {
             correct += 1;
         }
     }
